# docs/swagger.yaml
openapi: 3.0.3
security:
  - BearerAuth: []
info:
  title: AI-Persona Backend API
  version: 1.0.0
  description: |
    # AI-Persona SaaS Backend API - Core Authentication & User Management

    ## **Overview**
    This API provides enterprise-grade authentication and user management for the AI-Persona SaaS platform. The core module is designed for production environments with advanced security features and compliance.

    ## **Core Features**
    - **Authentication**: Local (email/password) and Google OAuth authentication
    - **Account Lifecycle**: PENDING_VERIFY → ACTIVE → DEACTIVATED → PENDING_DELETION
    - **Email Verification**: 24-hour TTL verification tokens with resend capability
    - **Password Reset**: Secure token-based password reset with 1-hour TTL
    - **Session Management**: Device tracking, session revocation, and token rotation
    - **Account Lockout**: Automatic lockout after failed login attempts (5 attempts → 15 min)
    - **GDPR Compliance**: Account deletion with 30-day grace period
    - **Webhook Integration**: External service integration for persona trait updates

    ## **Security Architecture**
    - **JWT Security**: RSA-256 signing with key rotation and JWKS endpoint
    - **Password Security**: Strength validation + HIBP breach detection
    - **Account Protection**: Rate limiting, lockout, session management
    - **Input Validation**: Comprehensive sanitization and validation
    - **Audit Trail**: Complete event logging with IP, user agent, trace IDs

...
    ```
...

...

...

...

...
servers:
...
paths:
...
  /api/auth/sessions:
    get:
      tags:
        - Authentication
      summary: Get user sessions
      description: Get all active sessions for the authenticated user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User sessions retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/sessions/{sessionId}:
    delete:
      tags:
        - Authentication
      summary: Revoke session
      description: Revoke a specific user session
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
          description: Session ID to revoke
      responses:
        "200":
          description: Session revoked successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/personas/{id}/chat:
    post:
      tags:
        - Personas
      summary: Send message to persona
      description: Send a message to a persona and get a response via webhook
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Persona ID
          schema:
            type: string
            example: "persona_id_here"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  minLength: 1
                  maxLength: 10000
                  example: "Hello, how are you today?"
                conversationId:
                  type: string
                  format: uuid
                  description: Optional conversation ID to continue existing conversation
                  example: "conversation_id_here"
                fileId:
                  type: string
                  description: Optional file ID for file attachment
                  example: "file_id_here"
      responses:
        "200":
          description: Message sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Message sent successfully"
                  data:
                    type: object
                    properties:
                      reply:
                        type: string
                        example: "Hello! I'm doing well, thank you for asking."
                      conversationId:
                        type: string
                        example: "conversation_id_here"
                      messageId:
                        type: string
                        example: "message_id_here"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "502":
          $ref: "#/components/responses/BadGateway"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # Chat Session Management
  /api/chat-sessions:
    get:
      tags:
        - Chat Sessions
      summary: Get user's chat sessions
      description: |
        Get a list of chat sessions for the authenticated user.
        Supports filtering by status and pagination.
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by session status
          required: false
          schema:
            type: string
            enum: [ACTIVE, COMPLETED, FAILED, TIMEOUT, CANCELLED]
            example: "ACTIVE"
        - name: limit
          in: query
          description: Number of sessions to return (max 100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
            example: 50
        - name: offset
          in: query
          description: Number of sessions to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
      responses:
        "200":
          description: Chat sessions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "User chat sessions retrieved successfully"
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ChatSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/chat-sessions/{sessionId}:
    get:
      tags:
        - Chat Sessions
      summary: Get specific chat session
      description: |
        Get details of a specific chat session including all messages.
        Users can only access their own sessions.
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: Chat session ID
          schema:
            type: string
            example: "abc123def456"
      responses:
        "200":
          description: Chat session retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Chat session retrieved successfully"
                  data:
                    $ref: "#/components/schemas/ChatSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    delete:
      tags:
        - Chat Sessions
      summary: Delete chat session
      description: |
        Permanently delete a chat session and all associated messages.
        Users can only delete their own sessions.
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: Chat session ID to delete
          schema:
            type: string
            example: "abc123def456"
      responses:
        "200":
          description: Chat session deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Chat session deleted successfully"
                  data:
                    type: object
                    properties:
                      sessionId:
                        type: string
                        example: "abc123def456"
                      deletedAt:
                        type: string
                        format: date-time
                        example: "2024-01-15T10:30:00Z"
                      messageCount:
                        type: integer
                        example: 5
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Error message"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "Invalid email format"

    # Authentication Response Schemas
...
    SessionsResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Sessions retrieved successfully"
        data:
          type: object
          properties:
            sessions:
              type: array
              items:
                $ref: "#/components/schemas/Session"

    # User Management Response Schemas
...
    Session:
      type: object
      properties:
        id:
          type: string
          example: "session_id_here"
        deviceId:
          type: string
          example: "device_id_here"
        userAgent:
          type: string
          example: "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36"
        ipAddress:
          type: string
          example: "192.168.1.1"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        expiresAt:
          type: string
          format: date-time
          example: "2024-01-22T10:30:00Z"
...
    Persona:
      type: object
      properties:
        id:
          type: string
          example: "persona_id_here"
        name:
          type: string
          example: "AI Assistant"
        description:
          type: string
          example: "A helpful AI assistant for general questions"
        avatarUrl:
          type: string
          format: uri
          example: "https://example.com/avatar.jpg"
        isActive:
          type: boolean
          example: true
        isAvailable:
          type: boolean
          example: true
        isFavourited:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        _count:
          type: object
          properties:
            conversations:
              type: integer
              example: 5
            messages:
              type: integer
              example: 25
    Conversation:
      type: object
      properties:
        id:
          type: string
          example: "conversation_id_here"
        title:
          type: string
          example: "Chat with AI Assistant"
        persona:
          type: object
          properties:
            id:
              type: string
              example: "persona_id_here"
            name:
              type: string
              example: "AI Assistant"
            avatarUrl:
              type: string
              format: uri
              example: "https://example.com/avatar.jpg"
        owner:
          type: object
          properties:
            id:
              type: string
              example: "user_id_here"
            name:
              type: string
              example: "John Doe"
            email:
              type: string
              format: email
              example: "john@example.com"
        visibility:
          type: string
          enum: [PRIVATE, SHARED]
          example: "PRIVATE"
        archivedAt:
          type: string
          format: date-time
          nullable: true
          example: null
        lastMessage:
          type: string
          example: "Hello! How can I help you today?"
        messageCount:
          type: integer
          example: 10
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    ChatSession:
      type: object
      properties:
        id:
          type: string
          example: "session_123"
        sessionId:
          type: string
          example: "abc123def456"
        conversationId:
          type: string
          example: "conv_456"
        personaId:
          type: string
          example: "persona_789"
        userId:
          type: string
          example: "user_123"
        status:
          type: string
          enum: [ACTIVE, COMPLETED, FAILED, TIMEOUT, CANCELLED]
          example: "COMPLETED"
        startedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:00:00Z"
        endedAt:
          type: string
          format: date-time
          nullable: true
          example: "2024-01-15T10:05:00Z"
        lastActivityAt:
          type: string
          format: date-time
          example: "2024-01-15T10:05:00Z"
        metadata:
          type: object
          nullable: true
          properties:
            userAgent:
              type: string
              example: "Mozilla/5.0..."
            ipAddress:
              type: string
              example: "192.168.1.1"
            deviceInfo:
              type: string
              example: "Chrome/120.0.0.0"
        errorMessage:
          type: string
          nullable: true
          example: "Webhook timeout"
        conversation:
          $ref: "#/components/schemas/Conversation"
        persona:
          $ref: "#/components/schemas/Persona"
        user:
          type: object
          properties:
            id:
              type: string
              example: "user_123"
            email:
              type: string
              example: "user@example.com"
            name:
              type: string
              example: "John Doe"
        messages:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: "msg_123"
              content:
                type: string
                example: "Hello, how are you?"
              role:
                type: string
                enum: [USER, ASSISTANT]
                example: "USER"
              createdAt:
                type: string
                format: date-time
                example: "2024-01-15T10:01:00Z"
      required:
        - id
        - sessionId
        - conversationId
        - personaId
        - userId
        - status
        - startedAt
        - lastActivityAt

  # Common Response Definitions
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Bad request"
              errors:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: "email"
                    message:
                      type: string
                      example: "Invalid email format"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Authentication required"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Insufficient permissions"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Resource not found"
...
    TooManyRequests:
      description: Too Many Requests
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Too many requests, please try again later"
    BadGateway:
      description: Bad Gateway
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Failed to get response from persona"
    ServiceUnavailable:
      description: Service Unavailable
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Persona is temporarily unavailable"